StartVote func created, voter tidy
